+++ lib/sync_engine_shim_for_ndk.dart
@@ -3,6 +3,7 @@ library;
 
 export 'src/entities/relay_filter_sync_state.dart';
 export 'src/entities/relay_knowledge.dart';
+export 'src/entities/sync_auth_error.dart';
 export 'src/entities/sync_engine_status.dart';
 export 'src/entities/sync_handle.dart';
 export 'src/entities/sync_progress.dart';